netops-automation-hub/requirements.txt
# SSH & Automation
netmiko==4.6.0
nornir==3.5.0
nornir-netmiko==1.0.1
nornir-utils==0.2.0

# API
fastapi==0.136.0
uvicorn==0.44.0

# Task Queue
celery==5.6.3
redis==7.4.0

# Database
psycopg2-binary==2.9.11
sqlalchemy==2.0.49

# Dashboard
streamlit==1.56.0

# Templating & Git
jinja2==3.1.6
gitpython==3.1.46

# Testing
pytest==9.0.3
httpx==0.28.1

# Utilities
python-dotenv==1.2.2
rich==15.0.0
